one more size_t fix, pixel error correction for alsa logo
git-svn-id: https://kradio.svn.sourceforge.net/svnroot/kradio/trunk@340 1c7eebff-f921-0410-8e14-d196b13b6ee1

+++ kradio3/plugins/timeshifter/timeshifter.cpp
@@ -204,7 +204,7 @@ bool TimeShifter::pausePlayback(SoundStreamID id)
 }
 
 
-unsigned TimeShifter::writeMetaDataToBuffer(const SoundMetaData &md, char *buffer, size_t buffer_size)
+size_t TimeShifter::writeMetaDataToBuffer(const SoundMetaData &md, char *buffer, size_t buffer_size)
 {
     Q_UINT64 pos     = md.position();
     time_t   abs     = md.absoluteTimestamp();
